EnSight readers: use double precision for time steps
(cherry picked from commit c6fd86fb5e13dcbd9b6c6646c77454b61c7eb0a7)

+++ IO/EnSight/vtkGenericEnSightReader.cxx
@@ -285,7 +285,7 @@ int vtkGenericEnSightReader::RequestData(vtkInformation* vtkNotUsed(request),
 }
 
 //------------------------------------------------------------------------------
-void vtkGenericEnSightReader::SetTimeValue(float value)
+void vtkGenericEnSightReader::SetTimeValue(double value)
 {
   vtkDebugMacro(<< this->GetClassName() << " (" << this << "): setting TimeValue to " << value);
   if (this->TimeValue != value)
